feat(driverbase): make it easier to bulk register info codes
