+++ driverbase/shared_assert.go
@@ -1,0 +1,48 @@
+// Copyright (c) 2025 ADBC Drivers Contributors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//         http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build assert
+
+package driverbase
+
+import (
+	"errors"
+	"fmt"
+
+	"github.com/apache/arrow-adbc/go/adbc"
+)
+
+func (sh *Shared[T]) Close() error {
+	if !sh.mu.TryLock() {
+		return adbc.Error{
+			Code: adbc.StatusInvalidState,
+			Msg:  fmt.Sprintf("[driverbase] Shared[%T].Close: resource is currently in use", *new(T)),
+		}
+	}
+	defer sh.mu.Unlock()
+
+	if sh.handle == nil {
+		return nil
+	}
+
+	if err := sh.closer.Close(); err != nil {
+		return errors.Join(adbc.Error{
+			Code: adbc.StatusInternal,
+			Msg:  fmt.Sprintf("[driverbase] Shared[%T].Close: failed to close resource: %s", *new(T), err),
+		}, err)
+	}
+	sh.handle = nil
+	sh.closer = nil
+	return nil
+}
